move test data and setup fixtures to test files
