+++ tests/bgp/test_bgp_aggregate_address.py
@@ -20,14 +20,17 @@ import pytest
 # Functions
 from bgp_bbr_helpers import config_bbr_by_gcu, get_bbr_default_state, is_bbr_enabled
 
-from bgp_aggregate_helpers import (  # noqa: F401
+from bgp_aggregate_helpers import (
+    BGP_AGGREGATE_ADDRESS,
     AggregateCfg,
+    dump_db,
     gcu_add_aggregate,
+    gcu_add_placeholder_aggregate,
     gcu_remove_aggregate,
-    setup_teardown,
     verify_bgp_aggregate_consistence,
     verify_bgp_aggregate_cleanup,
 )
+from tests.common.gcu_utils import create_checkpoint, rollback_or_reload, delete_checkpoint
 
 logger = logging.getLogger(__name__)
 
@@ -36,12 +39,25 @@ pytestmark = [
     pytest.mark.topology("m1"),
 ]
 
-# ---- Constants ----
-CONSTANTS_FILE = "/etc/sonic/constants.yml"
+
+@pytest.fixture(scope="module", autouse=True)
+def setup_teardown(duthost):
+    """Create checkpoint before tests, rollback after."""
+    create_checkpoint(duthost)
+    default_aggregates = dump_db(duthost, "CONFIG_DB", BGP_AGGREGATE_ADDRESS)
+    if not default_aggregates:
+        gcu_add_placeholder_aggregate(duthost, PLACEHOLDER_PREFIX)
+    yield
+    try:
+        rollback_or_reload(duthost, fail_on_rollback_error=False)
+    finally:
+        delete_checkpoint(duthost)
+
 
 # Aggregate prefixes
 AGGR_V4 = "172.16.51.0/24"
 AGGR_V6 = "2000:172:16:50::/64"
+PLACEHOLDER_PREFIX = "192.0.2.0/32"
 
 
 # Test with parameters Combination
